clean event layout
